backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy>=2.0.25
psycopg2-binary>=2.9.9
alembic==1.12.1
pydantic>=2.7.0
python-dotenv==1.0.0
google-generativeai==0.8.5
google-api-python-client>=2.0.0
openai>=1.0.0
python-multipart==0.0.6
fastapi-cors==0.0.6

# Authentication dependencies
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# LangChain dependencies for conversation history
langchain>=0.1.0
langchain-community>=0.0.20
langchain-core>=0.1.0
langchain-text-splitters>=0.0.1
langsmith>=0.0.1
